nft, rule: Support row expressions
The nft configuration supports a very large set of expressions.
In order to allow callers to use expressions which are not part of the
go-nft schema, a new field was added to the Expression struct:
`RowData`.

The field is mutually exclusive with the other fields and expects a
slice of bytes.

Signed-off-by: Edward Haas <edwardh@redhat.com>