automation: Support the option to disable IPv6 in the container
When IPv6 is available on the host, the `apk` application in the
container attempts to communicate using IPv6 which is not always
possible (e.g. due to the lack of support from the ISP).

Allow the test runner to disable IPv6 in the container by setting an
environment variable: `DISABLE_IPV6_IN_CONTAINER=1`

Signed-off-by: Edward Haas <edwardh@redhat.com>