rule: Support specifying a rule index
When adding rules, the index is used to determine the order in which the
new rule should be placed where multiple ones exist in a chain.

The first rule should not use an explicit index value, only the
following ones should add (append) or insert (prepend) rules with an
index.

An example will follow in the integration tests (in a following commit).

Signed-off-by: Edward Haas <edwardh@redhat.com>