tests: Introduce a no-mac-spoofing implementation example
The no-mac-spoofing example follows the configuration that libvirt is
applying when set with the no-mac-spoofing filter.

It is presented in the integration tests to present how to use `nft` API
to apply a configuration.
The example follows two formats, one that builds the configuration using
helper construct functions and the second that builds the spec directly.

Signed-off-by: Edward Haas <edwardh@redhat.com>