tests: Use testify require package instead of assert
The testify assertion library provides two packages for assertions. One
that will notify on the failure but will continue with other assertions
in the same test (assert package) and another that will stop the test on
the first failure and will continue to the next test (require package).

This behavior is confusing, as many developers would expect `assert` to
mean "stop here if it fails".

Therefore, the following change aliases the `require` package with the
`assert` name.
In case there will be a need to assert and still continue on with the
test, it will be considered an exception and a proper alias will be
given for the original assert package.

Signed-off-by: Edward Haas <edwardh@redhat.com>