exec, cgo: Add cgo binding to libnftables
By using libnftables directly, there is no need to exec a new process.
For the user perspective, it should look the same, as the details of how to
apply/read the nftables configuration is abstracted away.

exec or nftlib version can be selected at runtime as illustrated on the
following example:

```
import (
        "github.com/networkplumbing/go-nft/nft"
        nftlib "github.com/networkplumbing/go-nft/nft/lib"
)

nft.ReadConfig()
nftlib.ReadConfig()
```

Signed-off-by: Quique Llorente <ellorent@redhat.com>