tests, example: Add (simple) masquerade test
This change includes a skeleton example setup function which includes
common operations expected from the test: Apply the configuration, read
the new configuration and validate it corresponds to what has been applied
originally.

In addition, NAT rules require a dedicated kmod (`nft_masq`) which needs
to be loaded in the host. Although it loads lazily by default, in the
container used for the tests it cannot. Therefore, it is loaded before
the test container is launched.

Signed-off-by: Edward Haas <edwardh@redhat.com>