exec: Use a temp file to pass data to the `nft` binary
Using a buffer to pass the nftables configuration to `stdin` has been
unintentionally blocked by one of the latest SELinux policies [1].

While the policy issue is fixed, this change introduces an alternative
way to pass the configuration data using a temporary file.

The workaround can be removed once the SELinux policy is fixed.

[1] https://github.com/containers/container-selinux/issues/147

Signed-off-by: Edward Haas <edwardh@redhat.com>