automation, integ tests: Fix false-negative results
The integration tests run commands to execute two test runs: One that
unitizes the exec method and the second to utilize the lib method.

In case the first fails and the second passes, the overall result is a
pass, which is obviously wrong.

The script for running the tests has been patched to fail once an error
is detected.

Signed-off-by: Edward Haas <edwardh@redhat.com>