automation: Run integration tests serially
The integration tests are modifying the kernel iftables.
In order to allow tests to run independently, on each test a teardown
cleanup is executed. However, when tests run in parallel they will
collide and the cleanup is not helpful.

This change makes sure the tests are executed serially.
The parallel execution occurs due to the default parallel build of
multiple packages (and their run) when using `go test`.

It may be cleaner to run each test in its own network namespace,
allowing the test to run independently to how one may execute them.
But this is left for follow up improvements.

Signed-off-by: Edward Haas <edwardh@redhat.com>