automation: Load extra kernel-modules to support CentOS Stream 9
When running the tests on CentOS Stream 9, there is a need to load two
new kmods on the host that runs the tests (as the container in which the
tests are running has no permission to load a kmod).

To preserve the ability to run on older hosts, the `nft_meta_bridge`
kmod is not strictly required (as it is not available on older kernels).

Signed-off-by: Edward Haas <edwardh@redhat.com>