Adding handlers to the table and chain
Adding handlers to nftables structs and adjusted the test comparison
function to remove them.

Having the handlers in place when retrieving the configuration can be
useful to execute the nft binary directly with the retrieved handlers.
The rules already include the handle, this adds support for handlers in
tables and chains.

Signed-off-by: Tobias Buner <buner@anapaya.net>